[CI] fix
- Added phpcs
- Fixed CodeClimate
- PHPCS fixes
- Login refresh fix
- Added markdownlint config

+++ scripts/ci/install.sh
@@ -5,15 +5,19 @@ export BDD_PWD="$(openssl rand -base64 32)"
 export PATH=~/.composer/vendor/bin/:$PATH
 set -x
 set -e
-sudo /etc/init.d/mysql stop || echo "mysql not stopped"
-sudo stop mysql-5.6 || echo "mysql-5.6 not stopped"
-sudo  mysqld_safe --skip-grant-tables &
-sleep 4
-sudo mysql -e "update mysql.user set authentication_string=PASSWORD('$BDD_PWD') where User='root' and host='%'; update mysql.user set plugin='mysql_native_password'; delete from mysql.user where User != 'root' OR host != '%'; FLUSH PRIVILEGES;"
-sudo kill -9 `sudo cat /var/lib/mysql/mysqld_safe.pid`
-sudo kill -9 `sudo cat /var/run/mysqld/mysqld.pid`
-sudo /etc/init.d/mysql restart
-sleep 4
+
+if [[ "$TRAVIS_OS_NAME" = "osx" ]]; then
+    mysql -u root -ptoor -e "SET PASSWORD FOR 'root'@'localhost' = PASSWORD('$BDD_PWD'); FLUSH PRIVILEGES;"
+else
+    sudo /etc/init.d/mysql stop
+    sudo mysqld_safe --skip-grant-tables --user=root &
+    sleep 4
+    mysql -u root -e "update mysql.user set authentication_string=PASSWORD('$BDD_PWD') where User='root' and host='%'; update mysql.user set plugin='mysql_native_password'; delete from mysql.user where User != 'root' OR host != '%'; FLUSH PRIVILEGES;"
+    sudo kill -9 `sudo cat /var/lib/mysql/mysqld_safe.pid`
+    sudo kill -9 `sudo cat /var/run/mysqld/mysqld.pid`
+    sudo /etc/init.d/mysql start
+    sleep 4
+fi;
 
 mysql -uroot -p"$BDD_PWD" -e "CREATE DATABASE IF NOT EXISTS accountmanager CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci;"
 mysql -uroot -p"$BDD_PWD" accountmanager < $ME/../../sql/structure.sql
@@ -35,8 +39,14 @@ wget https://raw.githubusercontent.com/phpmyadmin/phpmyadmin/master/libraries/cl
 
 PMALICENSEURL="https://raw.githubusercontent.com/phpmyadmin/phpmyadmin/master/LICENSE"
 
-sed -i 's/PhpMyAdmin/AccountManager/g' $ME/../../src/Twig/*.php
-sed -i 's/PhpMyAdmin/AccountManager/g' $ME/../../src/Twig/I18n/*.php
-# Who care about vim :)
-sed -i "s,vim: set expandtab sw=4 ts=4 sts=4:,LICENSE : $PMALICENSEURL,g" $ME/../../src/Twig/*.php
-sed -i "s,vim: set expandtab sw=4 ts=4 sts=4:,LICENSE : $PMALICENSEURL,g" $ME/../../src/Twig/I18n/*.php
+if [[ "$TRAVIS_OS_NAME" = "osx" ]]; then
+    find "$ME/../../src/Twig" -type f -name '*.php' -exec sed -i '' -e 's,PhpMyAdmin,AccountManager,g' {} \;
+    # Who care about vim :)
+    find "$ME/../../src/Twig" -type f -name '*.php' -exec sed -i '' -e "s,vim: set expandtab sw=4 ts=4 sts=4:,LICENSE : $PMALICENSEURL,g" {} \;
+else
+    find "$ME/../../src/Twig" -type f -name '*.php' -exec sed -i 's,PhpMyAdmin,AccountManager,g' {} +
+    # Who care about vim :)
+    find "$ME/../../src/Twig" -type f -name '*.php' -exec sed -i "s,vim: set expandtab sw=4 ts=4 sts=4:,LICENSE : $PMALICENSEURL,g" {} +
+fi;
+
+
